Pull RewriteComparison.ComparisonOperator as top level class
